feat: 스크롤 내리면 secondline 숨기기
